Added a Kotlin folder with some basic examples and provided folder clean up
