+++ C/Bracket problem from hackerank
