+++ Kotlin/Hello World.kt
@@ -1,0 +1,4 @@
+/*fun main(args: Array<String>) {
+    println("Hello World!")
+}
+*/
