+++ PHP7/abstract_class.php
