+++ PHP7/compare_objects.php
